SPECTRA Auto ML Updated/requirements.txt
torch>=2.1.0
transformers>=4.44.0
datasets>=2.20.0
accelerate>=0.34.0
trl>=0.9.6
peft>=0.11.1
evaluate>=0.4.2
scikit-learn>=1.3.2
rich>=13.7.0
pyyaml>=6.0.2
jsonschema>=4.22.0
tqdm>=4.66.4

# AutoML dependencies
optuna>=3.3.0
scipy>=1.11.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.16.0
